refactor: modularize modules types
