+++ src/modules/authentication/types/index.js
@@ -1,0 +1,6 @@
+import { shape, bool, string } from 'prop-types'
+
+export const alertType = shape({
+  showAlert: bool.isRequired,
+  message: string
+}).isRequired
