+++ src/modules/rideRequest/containers/RideList.js
@@ -4,7 +4,8 @@ import PropTypes from 'prop-types'
 import { ListView } from 'react-native'
 import { fetchAllRides } from '../../../redux/actions'
 import { connect } from 'react-redux'
-import { Ride, ridePropTypes } from '../components/Ride'
+import { Ride } from '../components/Ride'
+import { ridePropTypes } from '../types/index'
 import { onNavigatorEvent } from '../../../navigation/NavBar'
 import styles from './styles/RideListStyles'
